added SimpleDataFeed example
