+++ templates/contracts/SimpleDataFeed.sol
@@ -1,0 +1,8 @@
+contract SimpleDataFeed {
+  // note: best used in conjunction with /examples/SimpleDataFeed
+  uint lastPrice;
+
+  function update(uint newPrice) {
+    lastPrice = newPrice;
+  }
+}
