+++ templates/views/partials/Contract.hbs
@@ -1,0 +1,14 @@
+{{>ContractJS}}
+{{>DeveloperStatus}}
+{{>ContractStatus}}
+
+{{#if contractIsCompiled }}
+      <div class="col-md-9" id="functionsDiv">
+        {{>WalletDiv}}
+        {{>ContractFunctions}}
+        {{>WalletModal}}
+     </div>
+<body onload="afterTX()">
+{{/if}}
+
+
